+++ R/net.mod.infection.R
@@ -56,10 +56,11 @@ infection.net <- function(dat, at) {
   if (nElig > 0 && nElig < nActive) {
 
     # Get discordant edgelist
-    del <- discord_edgelist(dat, at, network = 1)
+    del_list <- lapply(seq_len(dat$num.nw), discord_edgelist, dat = dat, at = at, include.network = TRUE)
+    del <- dplyr::bind_rows(del_list)
 
     # If some discordant edges, then proceed
-    if (!(is.null(del))) {
+    if (NROW(del) > 0) {
 
       # Infection duration to at
       del$infDur <- at - infTime[del$inf]
@@ -178,10 +179,11 @@ infection.2g.net <- function(dat, at) {
   if (nElig > 0 && nElig < nActive) {
 
     # Get discordant edgelist
-    del <- discord_edgelist(dat, at, network = 1)
+    del_list <- lapply(seq_len(dat$num.nw), discord_edgelist, dat = dat, at = at, include.network = TRUE)
+    del <- dplyr::bind_rows(del_list)
 
     # If some discordant edges, then proceed
-    if (!(is.null(del))) {
+    if (NROW(del) > 0) {
 
       # Infection duration to at
       del$infDur <- at - infTime[del$inf]
@@ -261,6 +263,8 @@ infection.2g.net <- function(dat, at) {
 #'        the current edgelist from. Default of \code{network = 1}.
 #' @param infstat Character vector of disease status values that are considered
 #'        infectious, defining the SI pairs.
+#' @param include.network Should the \code{network} value be included as the
+#'        final column of the discordant edgelist?
 #'
 #' @details
 #' This internal function works within the parent \code{\link{infection.net}}
@@ -289,7 +293,7 @@ infection.2g.net <- function(dat, at) {
 #' @export
 #' @keywords netMod internal
 #'
-discord_edgelist <- function(dat, at, network = 1, infstat = "i") {
+discord_edgelist <- function(dat, at, network = 1, infstat = "i", include.network = FALSE) {
 
   status <- get_attr(dat, "status")
   active <- get_attr(dat, "active")
@@ -318,6 +322,10 @@ discord_edgelist <- function(dat, at, network = 1, infstat = "i") {
         del <- NULL
       }
     }
+  }
+
+  if (NROW(del) > 0 && include.network == TRUE) {
+    del <- dplyr::bind_cols(del, network = network)
   }
 
   return(del)